fix: type conversion errors
Essentially this means marking previsouly implicit casts with explicit
ones. Also fixing the type of literal values. In some cases the used
types were refactored adequate ones.

+++ src/arch/armv8/armv8-a/psci.c
@@ -109,11 +109,11 @@ int32_t psci_cpu_suspend(uint32_t power_state, unsigned long entrypoint,
                         unsigned long context_id)
 {
 
-    return smc_call(PSCI_CPU_SUSPEND, power_state, entrypoint, context_id, NULL);
+    return (int32_t)smc_call(PSCI_CPU_SUSPEND, power_state, entrypoint, context_id, NULL);
 }
 
 int32_t psci_cpu_on(unsigned long target_cpu, unsigned long entrypoint, 
                     unsigned long context_id)
 {
-    return smc_call(PSCI_CPU_ON, target_cpu, entrypoint, context_id, NULL);
+    return (int32_t)smc_call(PSCI_CPU_ON, target_cpu, entrypoint, context_id, NULL);
 }
