+++ src/arch/riscv/mem.c
@@ -27,8 +27,8 @@ static inline void as_map_physical_identity(struct addr_space *as) {
         size_t num_entries = ((top - base - 1) / lvl_size) + 1;
 
         paddr_t addr = base;
-        for (int j = 0; j < num_entries; j++) {
-            int index = pt_getpteindex_by_va(&as->pt, (vaddr_t)addr, lvl);
+        for (size_t j = 0; j < num_entries; j++) {
+            size_t index = pt_getpteindex_by_va(&as->pt, (vaddr_t)addr, lvl);
             pte_set(&pt[index], addr, PTE_SUPERPAGE, PTE_HYP_FLAGS);
             addr += lvl_size;
         }
@@ -54,7 +54,7 @@ bool mem_translate(struct addr_space *as, vaddr_t va, paddr_t *pa)
             break;  
         }
         pte = (pte_t*)pte_addr(pte);
-        int index = pt_getpteindex_by_va(&as->pt, va, i+1);
+        size_t index = pt_getpteindex_by_va(&as->pt, va, i+1);
         pte = &pte[index];
     }
     if (pte && pte_valid(pte)) {
