+++ src/arch/riscv/sbi.c
@@ -46,31 +46,32 @@
  */
 #define SBI_EXTID_BAO (0x08000ba0)
 
-static inline struct sbiret sbi_ecall(long eid, long fid, long a0, long a1,
-                                      long a2, long a3, long a4, long a5)
+static inline struct sbiret sbi_ecall(unsigned long eid, unsigned long fid, 
+    unsigned long a0, unsigned long a1, unsigned long a2, unsigned long a3, 
+    unsigned long a4, unsigned long a5)
 {
-    long register _a0 __asm__("a0") = a0;
-    long register _a1 __asm__("a1") = a1;
-    long register _a2 __asm__("a2") = a2;
-    long register _a3 __asm__("a3") = a3;
-    long register _a4 __asm__("a4") = a4;
-    long register _a5 __asm__("a5") = a5;
-    long register _a6 __asm__("a6") = fid;
-    long register _a7 __asm__("a7") = eid;
+    unsigned long register _a0 __asm__("a0") = a0;
+    unsigned long register _a1 __asm__("a1") = a1;
+    unsigned long register _a2 __asm__("a2") = a2;
+    unsigned long register _a3 __asm__("a3") = a3;
+    unsigned long register _a4 __asm__("a4") = a4;
+    unsigned long register _a5 __asm__("a5") = a5;
+    unsigned long register _a6 __asm__("a6") = fid;
+    unsigned long register _a7 __asm__("a7") = eid;
 
     __asm__ volatile("ecall"
                  : "+r"(_a0), "+r"(_a1)
                  : "r"(_a2), "r"(_a3), "r"(_a4), "r"(_a5), "r"(_a6), "r"(_a7)
                  : "memory");
 
-    struct sbiret ret = {.error = _a0, .value = _a1};
+    struct sbiret ret = {.error = (long) _a0, .value = (long)_a1};
 
     return ret;
 }
 
 void sbi_console_putchar(int ch)
 {
-    (void)sbi_ecall(0x1, 0, ch, 0, 0, 0, 0, 0);
+    (void)sbi_ecall(0x1, 0, (unsigned long)ch, 0, 0, 0, 0, 0);
 }
 
 struct sbiret sbi_get_spec_version(void)
@@ -87,7 +88,7 @@ struct sbiret sbi_get_impl_version(void)
     return sbi_ecall(SBI_EXTID_BASE, SBI_GET_SBI_IMPL_VERSION_FID, 0, 0, 0, 0,
                      0, 0);
 }
-struct sbiret sbi_probe_extension(long extension_id)
+struct sbiret sbi_probe_extension(unsigned long extension_id)
 {
     return sbi_ecall(SBI_EXTID_BASE, SBI_PROBE_EXTENSION_FID, extension_id, 0,
                      0, 0, 0, 0);
@@ -284,7 +285,7 @@ static struct sbiret sbi_base_handler(unsigned long fid)
             ret.value = 0;
             for (size_t i = 0; i < NUM_EXT; i++) {
                 if (ext_table[i] == extid) {
-                    ret.value = extid;
+                    ret.value = (long)extid;
                 }
             }
             break;
@@ -357,7 +358,7 @@ static struct sbiret sbi_hsm_start_handler() {
                 ret.error = SBI_ERR_FAILURE;
             } else {
                 vaddr_t start_addr = vcpu_readreg(cpu()->vcpu, REG_A1);
-                unsigned priv = vcpu_readreg(cpu()->vcpu, REG_A2);
+                unsigned priv = (unsigned)vcpu_readreg(cpu()->vcpu, REG_A2);
                 vcpu->arch.sbi_ctx.state = START_PENDING;
                 vcpu->arch.sbi_ctx.start_addr = start_addr;
                 vcpu->arch.sbi_ctx.priv = priv;
@@ -454,8 +455,8 @@ size_t sbi_vs_handler()
             ret.value = SBI_ERR_NOT_SUPPORTED;
     }
 
-    vcpu_writereg(cpu()->vcpu, REG_A0, ret.error);
-    vcpu_writereg(cpu()->vcpu, REG_A1, ret.value);
+    vcpu_writereg(cpu()->vcpu, REG_A0, (unsigned long)ret.error);
+    vcpu_writereg(cpu()->vcpu, REG_A1, (unsigned long)ret.value);
 
     return 4;
 }
