+++ src/core/ipc.c
@@ -14,8 +14,8 @@ enum {IPC_NOTIFY};
 
 union ipc_msg_data {
     struct {
-        uint8_t shmem_id;
-        uint8_t event_id;
+        uint32_t shmem_id;
+        uint32_t event_id;
     };
     uint64_t raw;
 };
@@ -63,10 +63,10 @@ static void ipc_handler(uint32_t event, uint64_t data){
 }
 CPU_MSG_HANDLER(ipc_handler, IPC_CPUMSG_ID)
 
-unsigned long ipc_hypercall(unsigned long ipc_id, unsigned long ipc_event,
+long int ipc_hypercall(unsigned long ipc_id, unsigned long ipc_event,
                                                 unsigned long arg2)
 {
-    unsigned long ret = -HC_E_SUCCESS;
+    long int ret = -HC_E_SUCCESS;
 
     UNUSED_ARG(arg2);
 
@@ -82,8 +82,8 @@ unsigned long ipc_hypercall(unsigned long ipc_id, unsigned long ipc_event,
         cpumap_t ipc_cpu_masters = shmem->cpu_masters & ~cpu()->vcpu->vm->cpus;
 
         union ipc_msg_data data = {
-            .shmem_id = cpu()->vcpu->vm->ipcs[ipc_id].shmem_id,
-            .event_id = ipc_event,
+            .shmem_id = (uint32_t)cpu()->vcpu->vm->ipcs[ipc_id].shmem_id,
+            .event_id = (uint32_t)ipc_event,
         };
         struct cpu_msg msg = {IPC_CPUMSG_ID, IPC_NOTIFY, data.raw};
